# src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Performance Orchestrator API",
    "version": "0.1.0",
    "description": "CKP performance orchestration and integration configuration. Obtain the session cookie and CSRF header from /session for mutations. Execution is explicitly enabled through startup configuration."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/runs": {
      "get": {
        "summary": "Filtered run history",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environment",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profileId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/execution/monitoring-sets": {
      "get": {
        "summary": "List",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/monitoring-sets/{id}": {
      "delete": {
        "summary": "Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/execution/runs/{id}/monitoring": {
      "get": {
        "summary": "View",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Save",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/runs/{id}/monitoring/query": {
      "post": {
        "summary": "Query",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/settings": {
      "get": {
        "summary": "Settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/profiles": {
      "get": {
        "summary": "Profiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/services/{service}/values": {
      "post": {
        "summary": "Values",
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/preparations/{id}": {
      "get": {
        "summary": "Progress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/plans": {
      "post": {
        "summary": "Prepare",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/runs/{id}/loads": {
      "get": {
        "summary": "Loads",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/runs/{id}/loads/review": {
      "post": {
        "summary": "Review load",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/runs/{id}/loads/{loadId}/start": {
      "post": {
        "summary": "Start load",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "loadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/runs": {
      "post": {
        "summary": "Run",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/execution/runs/{id}/recover": {
      "post": {
        "summary": "Recover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "4XX": {
            "description": "Validation, authentication or conflict error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "field",
          "message"
        ]
      }
    }
  }
}